+++ widgets/ForwardWidgets-huangxd/widgets/danmu_auto.js
@@ -15,7 +15,7 @@
 WidgetMetadata = {
   id: "forward.auto.danmu2",
   title: "自动链接弹幕v2",
-  version: "2.0.3",
+  version: "2.0.4",
   requiredVersion: "0.0.2",
   description: "自动获取播放链接并从服务器获取弹幕【五折码：CHEAP.5;七折码：CHEAP】",
   author: "huangxd",
@@ -3143,7 +3143,7 @@ async function bahamutSearch(keyword) {
 
 async function getBahamutEpisodes(videoSn) {
   try {
-    const url = proxyUrl ? `http://127.0.0.1:5321/proxy?url=https://api.gamer.com.tw/anime/v1/video.php?videoSn=${videoSn}` : `https://api.gamer.com.tw/anime/v1/video.php?videoSn=${videoSn}`;
+    const url = `https://api.gamer.com.tw/anime/v1/video.php?videoSn=${videoSn}`;
     const resp = await Widget.http.get(url, {
       headers: {
         "Content-Type": "application/json",
@@ -3182,7 +3182,7 @@ async function fetchBahamutEpisodeDanmu(videoSn) {
   let danmus = [];
 
   try {
-    const url = proxyUrl ? `http://127.0.0.1:5321/proxy?url=https://api.gamer.com.tw/anime/v1/danmu.php?geo=TW%2CHK&videoSn=${videoSn}` : `https://api.gamer.com.tw/anime/v1/danmu.php?geo=TW%2CHK&videoSn=${videoSn}`;
+    const url = `https://api.gamer.com.tw/anime/v1/danmu.php?geo=TW%2CHK&videoSn=${videoSn}`;
     const resp = await Widget.http.get(url, {
       headers: {
         "Content-Type": "application/json",
@@ -3211,9 +3211,9 @@ function formatBahamutComments(items) {
   const positionToMode = { 0: 1, 1: 5, 2: 4 };
   return items.map(c => ({
     cid: Number(c.sn),
-    p: `${c.time.toFixed(2)},${positionToMode[c.position] || c.tp},${parseInt(c.color.slice(1), 16)},[bahamut]`,
+    p: `${Math.round(c.time / 10).toFixed(2)},${positionToMode[c.position] || c.tp},${parseInt(c.color.slice(1), 16)},[bahamut]`,
     m: simplized(c.text),
-    t: c.time
+    t: Math.round(c.time / 10)
   }));
 }
 
@@ -3228,7 +3228,7 @@ async function getBahamutComments(pid, progressCallback=null){
   log("info", `弹幕处理完成，共 ${formatted.length} 条`);
   // 输出前五条弹幕
   log("info", "Top 5 danmus:", JSON.stringify(formatted.slice(0, 5), null, 2));
-  return formatted;
+  return convertToDanmakuJson(formatted, "bahamut");
 }
 
 // =====================