+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-10-22 03:03:20
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: cc08f89dd6de93c313baf19ec5d5a702497e6e02
+
+## [ForwardWidgets-2kuai] - 2025-10-22 03:03:20
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 008108fc97646dd67bd9a83278d5ec527326c655
+
+## [ForwardWidgets-pack1r] - 2025-10-22 03:03:20
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: f11737fdaae2e6b47721c8485d6f8b7bfe612604
+
+
 ## [ForwardWidgets-huangxd] - 2025-10-21 02:58:55
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main